// client/src/types/siparis.ts
export interface SiparisData {
  Sec: boolean;
  SiparisID: number;
  SiparisNo: string;
  Tarih: string;
  FirmaID: number;
  PersonelID: number;
  SiparisDurumID: number;
  OdemeSartlari: string;
  Tasima: string;
  TeslimYeri: string;
  Aciklama: string;
  Personel: string;
  FirmaAdi: string;
  FirmaKodu: string;
  SiparisDurumu: string;
  SiparisDetayID: number;
  MalzemeID: number;
  BirimID: number;
  Miktari: number;
  BirimFiyati: number;
  Doviz: string;
  TerminTarihi: string;
  TerminHaftasi: number;
  SiparisDetayAciklama: string;
  BirimAdi: string;
  MalzemeKodu: string;
  MalzemeAdi: string;
  TestYontemStandarti: string;
  IlkGelisTarihi: string | null;
  IptalMiktari: number;
  IadeMiktari: number;
  GelenMiktar: number;
  KalanMiktar: number;
  Gecikme: number;
  MalzemeGrubu: string;
  MalzemeCinsi: string;
  MalzemeTuru: string;
  KDVOran: number;
  KDVTutar: number;
  KDVliTutar: number;
  TesisKodu: string;
  TesisAdi: string;
  Ambalaj: string;
  OdemeSekli: string;
  VadeGunu: number;
  MalzemeOlculeri: string;
  MalzemeOzelKodu1: string;
  MalzemeOzelKodu2: string;
  UreticiKodu: string;
  IlkGecikme: number;
  LME: number;
  Premium: number;
  TalepNo: string;
  SozlesmeBirimFiyati: number;
  SozLMEDonemAciklama: string;
  DosyaVarMi: boolean;
  TeklifNo: string;
  AmbalajNo: string;
  TalepTerminTarihi: string;
  TalepEden: string;
  BirimUzunluk: number;
  Uzunluk: number;
  KapamaDurumu: boolean;
  SiparisOzelKodu1: string;
  SiparisOzelKodu2: string;
  MalzemeTedarikciKodu: string;
  MalzemeTedarikciAdi: string;
  Tutari: number;
  GKTestVarMi: boolean;
  TalepKaynagi: string;
  TalepMasrafAciklamasi: string;
  Iskonto: number;
  IskontoTutar: number;
  SiparisBakiyeTutari: number;
  NumuneID: number | null;
  NumuneFormNo: string | null;
  Departman: string;
  TerminTarihiGunFarki: number;
  TerminAciklamasi: string;
  Marka: string;
  PozNo: string;
  IlkTerminTarihi: string;
  TalepEdenDepartman: string;
  SevkSekli: string;
  TalepNedeni: string;
  BirimMiktari: number;
  Adedi: number;
  Kesit: string;
  Tip: string;
  ListeFiyati: number;
  SiraNo: number;
  GTIPNo: string;
  SatisSiparisNo: string;
  Metin10: string;
  SonHareketTarihi: string;
}